Initial Commit
A Barebones IRC client. Right now it just spits the parameters field of
numeric replies at you, and allows you to type in IRC commands. It
autoreplies to PING and VERSION messages, and automatically registers
the user as `user_name' with password `very_secure_password'. The server
and port can be specified on the command line.