Channel tabs and /HELP
Each /JOINed channel gets its own tab on the left side of the screen.
Left clicking the channel's name will switch to it. I also added a /HELP
command that prints a help message. I also fixed a bug where the PART
message was missing the channel part when it was sent back the the GUI.
Finally, there should be fewer cases of zombie processes being left
behind if the program crashes.